refactor: reorganize file structure
- Move API services to api/ folder
  * ActivePortfolioApiService
  * AdventureSectionService
  * AdventureService
  * PortfolioService
- Move network config to network/ folder
  * RetrofitClient
  * AuthInterceptor
- Update package declarations and imports
