+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/api/ActivePortfolioApiService.kt
@@ -1,6 +1,5 @@
-package com.example.active_portfolio_mobile.data.remote.network
+package com.example.active_portfolio_mobile.data.remote.api
 
-import com.example.active_portfolio_mobile.data.remote.api.UserPublicApiService
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
