+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/api/PortfolioService.kt
@@ -1,4 +1,4 @@
-package com.example.active_portfolio_mobile.data.remote.network
+package com.example.active_portfolio_mobile.data.remote.api
 
 
 import com.example.active_portfolio_mobile.data.remote.dto.CreatePortfolioRequest
@@ -13,7 +13,6 @@ import retrofit2.http.Header
 import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Path
-import retrofit2.http.Query
 
 
 interface PortfolioService {
