+++ app/src/main/java/com/example/active_portfolio_mobile/viewModels/AdventureVM.kt
@@ -3,7 +3,7 @@ package com.example.active_portfolio_mobile.viewModels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.active_portfolio_mobile.data.remote.dto.Adventure
-import com.example.active_portfolio_mobile.data.remote.network.ActivePortfolioApi
+import com.example.active_portfolio_mobile.data.remote.api.ActivePortfolioApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
