+++ app/src/main/java/com/example/active_portfolio_mobile/viewModels/SinglePortfolio.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.active_portfolio_mobile.data.remote.dto.CreatePortfolioRequest
 import com.example.active_portfolio_mobile.data.remote.dto.Portfolio
 import com.example.active_portfolio_mobile.data.remote.dto.UpdatePortfolioRequest
-import com.example.active_portfolio_mobile.data.remote.network.ActivePortfolioApi
+import com.example.active_portfolio_mobile.data.remote.api.ActivePortfolioApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
